+++ server/Application/DTOs/FeedIntakeInsertDTO.cs
@@ -3,6 +3,6 @@ namespace Application.DTOs
     public class FeedIntakeInsertDTO
     {
         public string FeedId { get; set; }
-        public int ExpectedQuantity { get; set; }
+        public decimal ExpectedQuantity { get; set; }
     }
 }
